% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{new_db_pool}
\alias{new_db_pool}
\title{New DB Pool}
\usage{
new_db_pool(config)
}
\arguments{
\item{config}{A list of connection arguments consumed by \code{{pool}} and \code{{odbc}}}
}
\value{
A database connection pool
}
\description{
Creates a new database connection pool
}
